+++ R/zzz.R
@@ -11,6 +11,10 @@
 # been loaded.
 
 .onLoad <- function(libname, pkgname) {
+
+    # CRAN OMP THREAD LIMIT
+    Sys.setenv("OMP_THREAD_LIMIT" = 1)
+
     # This defines the model database
 
     # Prophet