+++ tests/testthat/test-tune_workflows.R
@@ -56,7 +56,11 @@ test_that("Tuning, arima_boost", {
             resamples = resample_spec,
             grid      = grid_spec,
             metrics   = metric_set(mae, mape, smape, mase, rmse, rsq),
-            control   = control_grid(verbose = FALSE, allow_par = TRUE)
+            control   = control_grid(
+                verbose   = FALSE,
+                allow_par = TRUE,
+                cores     = 2
+            )
         )